requirements.txt
streamlit
requests
Pillow
python-dotenv
openai
streamlit-audiorecorder
geopy
streamlit-js-eval
